Migrate content of RamPrimitiveTypes to RamTypes and Util
