+++ src/CompiledSouffle.h
@@ -24,7 +24,6 @@
 #include "souffle/IODirectives.h"
 #include "souffle/IOSystem.h"
 #include "souffle/ParallelUtils.h"
-#include "souffle/RamPrimitiveTypes.h"
 #include "souffle/RamTypes.h"
 #include "souffle/SignalHandler.h"
 #include "souffle/SouffleInterface.h"
