+++ src/Makefile.am
@@ -196,7 +196,6 @@ soufflepublic_HEADERS = \
         PiggyList.h                               \
         ProfileDatabase.h                         \
         ProfileEvent.h                            \
-	RamPrimitiveTypes.h                       \
         RamTypes.h                                \
         ReadStream.h                              \
         ReadStreamCSV.h                           \
