+++ src/Synthesiser.cpp
@@ -24,10 +24,10 @@
 #include "RamIndexAnalysis.h"
 #include "RamNode.h"
 #include "RamOperation.h"
-#include "RamPrimitiveTypes.h"
 #include "RamProgram.h"
 #include "RamRelation.h"
 #include "RamTranslationUnit.h"
+#include "RamTypes.h"
 #include "RamUtils.h"
 #include "RamVisitor.h"
 #include "RelationRepresentation.h"
