+++ src/Util.h
@@ -103,23 +103,52 @@ inline unsigned long __builtin_ctzll(unsigned long long value) {
 }
 #endif
 
-/**
- * Converts a string to a RamDomain (i.e. signed 32 bit or 64 bit number)
- */
-
 namespace souffle {
 
+/**
+ * Converts a string to a RamDomain
+ */
+inline RamDomain RamDomainFromString(const std::string& str, std::size_t* position = nullptr, int base = 10) {
+    RamDomain val;
 #if RAM_DOMAIN_SIZE == 64
-inline RamDomain stord(const std::string& str, std::size_t* pos = nullptr, int base = 10) {
-    return static_cast<RamDomain>(std::stoull(str, pos, base));
-}
-#elif RAM_DOMAIN_SIZE == 32
-inline RamDomain stord(const std::string& str, std::size_t* pos = nullptr, int base = 10) {
-    return static_cast<RamDomain>(std::stoul(str, pos, base));
-}
+    val = std::stoll(str, position, base);
 #else
-#error RAM Domain is neither 32bit nor 64bit
+    val = std::stoi(str, position, base);
 #endif
+    return static_cast<RamDomain>(val);
+}
+
+/**
+ * Converts a string to a RamFloat
+ */
+inline RamFloat RamFloatFromString(const std::string& str, std::size_t* position = nullptr) {
+    RamFloat val;
+#if RAM_DOMAIN_SIZE == 64
+    val = std::stod(str, position);
+#else
+    val = std::stof(str, position);
+#endif
+    return static_cast<RamFloat>(val);
+}
+
+/**
+ * Converts a string to a RamUnsigned
+ */
+inline RamUnsigned RamUnsignedFromString(
+        const std::string& str, std::size_t* position = nullptr, int base = 10) {
+    RamUnsigned val;
+#if RAM_DOMAIN_SIZE == 64
+    val = std::stoul(str, position, base);
+#else
+    val = std::stoull(str, position, base);
+#endif
+    return static_cast<RamUnsigned>(val);
+}
+
+// This needs to be removed.
+inline RamDomain stord(const std::string& str, std::size_t* pos = nullptr, int base = 10) {
+    return RamDomainFromString(str, pos, base);
+}
 
 /**
  * Check whether a string is a sequence of digits
